Move shadow constants to separate file
