+++ src/components/Card/Card.story.js
@@ -4,7 +4,7 @@ import { withInfo } from '@storybook/addon-info';
 import { action } from '@storybook/addon-actions';
 
 import withTests from '../../util/withTests';
-import { KILO } from '../../util/sizes';
+import { KILO } from '../../styles/sizes';
 import Card, { CardHeader, CardFooter } from '.';
 import Heading from '../Heading';
 import Text from '../Text';
