+++ src/components/List/List.js
@@ -4,7 +4,7 @@ import styled, { css } from 'react-emotion';
 import { childrenPropType } from '../../util/shared-prop-types';
 import HtmlElement from '../HtmlElement/HtmlElement';
 import { textMega, textKilo } from '../../styles/style-helpers';
-import { BIT, BYTE, KILO, MEGA } from '../../util/sizes';
+import { BIT, BYTE, KILO, MEGA } from '../../styles/sizes';
 
 const baseStyles = ({ theme }) => css`
   label: list;
