+++ src/components/LoadingBar/LoadingBar.js
@@ -5,7 +5,7 @@ import styled, { css } from 'react-emotion';
 import { childrenPropType } from '../../util/shared-prop-types';
 import { textKilo, subtractUnit } from '../../styles/style-helpers';
 import { uniqueId } from '../../util/id';
-import { KILO, MEGA, GIGA } from '../../util/sizes';
+import { KILO, MEGA, GIGA } from '../../styles/sizes';
 
 const calculateSize = ({ theme, size }) => {
   const sizeMap = {
