+++ src/styles/shadows.js
@@ -1,0 +1,3 @@
+export const SINGLE = 'single';
+export const DOUBLE = 'double';
+export const TRIPLE = 'triple';
